AVRO-1846: C: change the refcount type from "int" to "long"
The refcount is currently stored in the same memory allocation, right
before the value; src/generic reads:
    void  *self = avro_malloc(instance_size + sizeof(volatile int));
    [...]
    volatile int  *refcount = (volatile int *) self;
    self = (char *) self + sizeof(volatile int);

On 64-bit architectures, however, sizeof(int) == 4, which means that
effectively all accesses to "self" from there onwards would be shifted
by 32-bits and thus unaligned. This is slow on many (most?)
architectures, and entirely broken (SIGBUS) on very few (e.g. SPARC-V9).

Change the datatype of the refcount to long, as sizeof(long) equals
size(void *) (at least on Unix), which should make this functional, fast
and memory efficient on both 32-bit and 64-bit architectures.

+++ lang/c/src/avro/basics.h
@@ -53,7 +53,7 @@ typedef enum avro_class_t avro_class_t;
 struct avro_obj_t {
 	avro_type_t type;
 	avro_class_t class_type;
-	volatile int  refcount;
+	volatile long  refcount;
 };
 
 #define avro_classof(obj)     ((obj)->class_type)
