+++ lang/c/src/generic.c
@@ -57,7 +57,7 @@ avro_generic_value_new(avro_value_iface_t *iface, avro_value_t *dest)
 	avro_generic_value_iface_t  *giface =
 	    container_of(iface, avro_generic_value_iface_t, parent);
 	size_t  instance_size = avro_value_instance_size(giface);
-	void  *self = avro_malloc(instance_size + sizeof(volatile int));
+	void  *self = avro_malloc(instance_size + sizeof(volatile long));
 	if (self == NULL) {
 		avro_set_error(strerror(ENOMEM));
 		dest->iface = NULL;
@@ -65,8 +65,8 @@ avro_generic_value_new(avro_value_iface_t *iface, avro_value_t *dest)
 		return ENOMEM;
 	}
 
-	volatile int  *refcount = (volatile int *) self;
-	self = (char *) self + sizeof(volatile int);
+	volatile long  *refcount = (volatile long *) self;
+	self = (char *) self + sizeof(volatile long);
 
 	*refcount = 1;
 	rval = avro_value_init(giface, self);
@@ -90,8 +90,8 @@ avro_generic_value_free(const avro_value_iface_t *iface, void *self)
 		    container_of(iface, avro_generic_value_iface_t, parent);
 		size_t  instance_size = avro_value_instance_size(giface);
 		avro_value_done(giface, self);
-		self = (char *) self - sizeof(volatile int);
-		avro_free(self, instance_size + sizeof(volatile int));
+		self = (char *) self - sizeof(volatile long);
+		avro_free(self, instance_size + sizeof(volatile long));
 	}
 }
 
@@ -102,7 +102,7 @@ avro_generic_value_incref(avro_value_t *value)
 	 * This only works if you pass in the top-level value.
 	 */
 
-	volatile int  *refcount = (volatile int *) ((char *) value->self - sizeof(volatile int));
+	volatile long  *refcount = (volatile long *) ((char *) value->self - sizeof(volatile long));
 	avro_refcount_inc(refcount);
 }
 
@@ -113,7 +113,7 @@ avro_generic_value_decref(avro_value_t *value)
 	 * This only works if you pass in the top-level value.
 	 */
 
-	volatile int  *refcount = (volatile int *) ((char *) value->self - sizeof(volatile int));
+	volatile long  *refcount = (volatile long *) ((char *) value->self - sizeof(volatile long));
 	if (avro_refcount_dec(refcount)) {
 		avro_generic_value_free(value->iface, value->self);
 	}
@@ -147,7 +147,7 @@ struct avro_generic_link_value_iface {
 	avro_generic_value_iface_t  parent;
 
 	/** The reference count for this interface. */
-	volatile int  refcount;
+	volatile long  refcount;
 
 	/** The schema for this interface. */
 	avro_schema_t  schema;
@@ -1910,7 +1910,7 @@ avro_generic_string_new_length(avro_value_t *value, const char *str, size_t size
 
 typedef struct avro_generic_array_value_iface {
 	avro_generic_value_iface_t  parent;
-	volatile int  refcount;
+	volatile long  refcount;
 	avro_schema_t  schema;
 	avro_generic_value_iface_t  *child_giface;
 } avro_generic_array_value_iface_t;
@@ -2162,7 +2162,7 @@ avro_generic_array_class(avro_schema_t schema, memoize_state_t *state)
 
 typedef struct avro_generic_enum_value_iface {
 	avro_generic_value_iface_t  parent;
-	volatile int  refcount;
+	volatile long  refcount;
 	avro_schema_t  schema;
 } avro_generic_enum_value_iface_t;
 
@@ -2335,7 +2335,7 @@ avro_generic_enum_class(avro_schema_t schema)
 
 typedef struct avro_generic_fixed_value_iface {
 	avro_generic_value_iface_t  parent;
-	volatile int  refcount;
+	volatile long  refcount;
 	avro_schema_t  schema;
 	size_t  data_size;
 } avro_generic_fixed_value_iface_t;
@@ -2545,7 +2545,7 @@ avro_generic_fixed_class(avro_schema_t schema)
 
 typedef struct avro_generic_map_value_iface {
 	avro_generic_value_iface_t  parent;
-	volatile int  refcount;
+	volatile long  refcount;
 	avro_schema_t  schema;
 	avro_generic_value_iface_t  *child_giface;
 } avro_generic_map_value_iface_t;
@@ -2833,7 +2833,7 @@ avro_generic_map_class(avro_schema_t schema, memoize_state_t *state)
 
 typedef struct avro_generic_record_value_iface {
 	avro_generic_value_iface_t  parent;
-	volatile int  refcount;
+	volatile long  refcount;
 	avro_schema_t  schema;
 
 	/** The total size of each value struct for this record. */
@@ -3209,7 +3209,7 @@ error:
 
 typedef struct avro_generic_union_value_iface {
 	avro_generic_value_iface_t  parent;
-	volatile int  refcount;
+	volatile long  refcount;
 	avro_schema_t  schema;
 
 	/** The total size of each value struct for this union. */
